pyproject.toml
[tool.poetry]
name = "vend-o-matic--pzc"
version = "0.1.0"
description = ""
authors = ["Paul Zimmerman-Clayton <paul@pzc.dev>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
beautifulsoup4 = "^4.11.1"

[tool.poetry.dev-dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
